Clean up of errors and traits
